Move nan functions to new module and remove leading underscore
